All set in Booking

+++ src/main/java/com/irctc/trainbookingsystem/dto/UserDto.java
@@ -9,7 +9,7 @@ public class UserDto {
 	private String lastName;
 	private long phoneNo;
 	private String email;
-	private int walletBalance;
+	private long walletBalance;
 	
 	public int getUserId() {
 		return userId;
@@ -53,10 +53,11 @@ public class UserDto {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public int getWalletBalance() {
+	public long getWalletBalance() {
 		return walletBalance;
 	}
-	public void setWalletBalance(int walletBalance) {
+	public void setWalletBalance(long walletBalance) {
 		this.walletBalance = walletBalance;
 	}
+	
 }
